+++ tenement-domain/src/main/java/com/tenement/domain/dto/RoomNewDetailsResp.java
@@ -55,7 +55,7 @@ public class RoomNewDetailsResp implements Serializable {
 
     private Byte finishType;
 
-    private Byte districtid;
+    private Integer districtid;
 
     private Byte heatingTyte;
